Remove unreachable and unnecessary code in gossip membership
In gossip, alive messages are stored in membership stores, and they also have logical timestamps.

When a peer learns of an alive message with a higher logical timestamp of its latest alive message (for a specific peer),
it enters learnExistingMembers() and tentatively updates the latest envelope and gossip message of what is store in the membership store.

However, for completeness, the code checks if the membership store already contains an earlier instance, and if not, populates the membership store
with a new instance. This population is not needed, because the code only enters the function if it already knows the peer, hence the membership
store contains an earlier instance.

This commit simply removes the unneeded code path where the instance is unknown despite the fact that we know about the peer already.

Change-Id: I86bc65d881a440c267275a697853a1ead087df18
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>