FABGW-8: Add a commit notifier for embedded Gateway
Add a notifier that enables the embedded Gateway to get the validation code for a specific transaction to be submitted to the orderer, once it is committed to the local peer's ledger. The notifier provides an abstraction of the way transaction commits are obtained to enable an alternative implementation (with a shared interface) to be developed for a standalone Gateway.

Commit notification is not yet wired into the Gateway runtime. This will occur in a follow-on change that enables the end-to-end submit transaction flow for the embedded Gateway.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>