FABGW-6 Implement Submit()
Use discovery to get peer endorsers and orderers in order to submit a transaction to the ledger

A number of TODO comments are in the code marking where additional logic will be added as part of this user story (or future story where specified). This includes adding retry logic according to endorsement layouts, configurable timeouts, updating stored channel membership in response to config updates.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>