nwo: add nwo.SigningIdentity and accessors
Test suites that exercise the gRPC API instead of the command line tools
were instantiating an MSP to obtain a signing identity. In order to
enable reuse of code and reduce the coupling of tests to the MSP and
BCCSP, this commit introduces a Signer that satisfies the `Sign` and
`Serialize` operations used when assembling protocol buffer messages.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>