Fix bug in snapshot request submission
This commit fixes a bug in the snapshot request path.
When the snapshot request is sent with the special value "blockNumber 0",
if a block commit is in progress, the derived value should be one higher
than the last committed block (i.e., equal to the currently processing block).
Instead in the code, it was a typo that made it always set to 1.

Also, add a wait in the discovery integration test for channel config
updates to be committed

cherry-picked commit 44abc6b26d964eab9adf83d17a58e4ceb96ed39c

Signed-off-by: manish <manish.sethi@gmail.com>