integration: remove helpers package
The two functions in the package are used to verify that docker images
exist and to generate a unique name for a docker network and containers
and each function had a single consumer.

This commit moves `UniqueName` into the `runner` package and the image
assertion block out of the build server and into the network
constructor.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>