integration: remove yaml network definitions
When `nwo` was first created, we expected the network model to be be
read from a file instead of being constructed at runtime so we included
an example that used that mechanism. We've since learned that we can get
a way with a small number of "standard" networks (with minor deviations)
and that the serialization is not required.

This change removes network definition fixtures from the integration
tests and removes the example from `nwo`.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>