Replace WaitGroups with channels and timeouts
A sync.WaitGroup does not have a timeout mechanism. If a test is told to
Wait on a WaitGroup and the required number of Done calls never get
made, the test will hang until the testing package finally kills it.
When that happens, we get some stack traces but we lose a lot of time
and a fair bit of information.

After hitting the 20m test time out in common/cluster three times in one
week, I decided it's time to add a proper timeout.

This commit switches from a WaitGroup to a channel, a sync.Once, and a
select on the channel close and a timeout condition. This should result
in test failures instead of termination from the testing package.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>