pkcs11: Add SKI to CKA_ID mapping for BCCSP
This adds support for finding PKCS#11 public and private key objects
that do not have a CKA_ID that matches the SKI of the public key.

The PKCS#11 provider can be configured with a map of hex-encoded subject
key identifiers (hashes) to CKA_ID values (strings) of the associated
objects. When the 'KeyIDs' configuration is present, this mapping is
used to locate key objects.

'AltID' is provided to match the behavior of Fabric 1.4 where key
identifiers are ignored and all SKIs are mapped to a single PKCS#11
CKA_ID. This should be considered deprecated but generally works as most
Fabric components only use a single signing identity.

If both 'AltID' and 'KeyIDs' elements are present, the 'AltID' value is
used when a mapping is not found in 'KeyIDs'.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>