Add GO111MODULE=on to command
Go modules are on by default now in the latest versions
of Go. Users who have cloned their projects onto their
GOPATH however will be unable to download the dependency
if they are GOM111ODULE is in `auto` or `off` mode. Adding
the GO111MODULE parameter to the command will ensure
all users are able to download the dependency.

Signed-off-by: Brett Logan <lindluni@github.com>