[FAB-18479] Log error if orderer can't forward SubmitRequest to Raft leader
If order cannot forward client transaction SubmitRequest to Raft leader,
it only logged it in a debug message. And since the error does
not get returned to the client either, it is difficult to identify
and troubleshoot.

This commit changes the message to an error in the orderer log.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>
(cherry picked from commit e8e39e6dfcceb14a4998d6ea64542d9d8a7b825b)