Use protoc-gen-go 1.3.3 for generating protos
A previous commit #2113 upgraded the protobuf in tools while the
protobuf in fabric dependencies is still 1.3.3. This upgrade was caused
by swagger inclusion.

In the current form, `make protos` causes to use upgraded version of
protoc-gen-go. This, in turn, requires to upgrade the protobuf lib in
the fabric dependancies -- which we could not until we resolve some
issues. See FAB-18363.

Signed-off-by: manish <manish.sethi@gmail.com>