[FAB-18508] ledger utility always outputs txNum
This patch fixes a bug reported in FAB-18508, and makes the ledger
troubleshooting utility output always the record "txNum".

The bug occurred when the target transaction is the first transaction of
a block (i.e. txNum == 0). This patch removes the `omitempty` attributes
from all the members of a record.

Note that this patch also changes the output when a key is missing in
either snapshot. For a missing snapshot, 'null' will be output instead
of an empty object ({}).

Signed-off-by: Taku Shimosawa <taku.shimosawa.uf@hitachi.com>