[FAB-11334] Scrubs partially constructed/deleted ledgers at peer init (#2754)
Unjoining a peer from a channel will mark the ledger with status UNDER_DELETION
and proceed with the ledger removal.  If the unjoin operation fails, the ledger
will include residue from the partial deletion, leaving the kvledger in a
questionable state.  This commit forces the peer initialization to scan for any
ledgers with an UNDER_DELETION status, removing partial ledgers from the peer.

Signed-off-by: Josh Kneubuhl <jkneubuh@us.ibm.com>