[FAB-18527] Discovery supports DisregardNamespacePolicy hint from client
This commit adds support for a boolean flag in the chaincode call: DisregardNamespacePolicy
Its role is to hint to discovery that the client wishes to ask about endorsement policies
of either state based endorsement policies of given keys, or of collection endorsement policies
that exist for given collection names.

In case no state based endorsement policies are given in the bespoken chaincode call,
and also no collection level endorsement policies of the given collection names are found,
specifying the DisregardNamespacePolicy hint for that chaincode call will result in an error,
as there is no endorsement policy to compute on.

Change-Id: I299afd40063a0cf1cee34d6034eaa02fe6fc6707
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>