Improve an error message in `InstallChaincode`
In case a 'chaincode already successfully installed' error occurs when
installing a chaincode, Fabric admins cannot to know the package ID.
In this case, if multiple chaincode packages with the same label are installed,
it is not possible to identify them later by using `queryinstalled`.

So, this patch adds the package ID to the error message in the case of
'chaincode already successfully installed'.

Signed-off-by: Tatsuya Sato <tatsuya.sato.so@hitachi.com>