Evaluate() error response for node chaincode
The gateway Evaluate() method is not returning the error response to the client for node chaincodes. The code path for node chaincode errors is different than for Go chaincodes (this is a known separate issue).
It turns out the endorser is handling this difference and populates the Response field of the ProposalResponse accordingly.  The gateway needs to use this field rather than the one in the ChaincodeAction structure.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>