Update Sphinx to v1.8.2
A recent commit updated the doc generation tool, sphinx, from v1.7.2 to v1.8.0.  However, this broke the ability to generate and review the docs from a forked branch.
This is due to a known bug in Sphinx which was fixed on v1.8.2 (https://github.com/sphinx-doc/sphinx/issues/5519).
This commit updates the dependency to this version.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>