ProposalResponse error should return error
A transaction proposal can fail for two reasons:
1) The chaincode return an error (status 500) response
2) Some infrastructure error occurs or error generated in the peer.

For case (1), the ProcessProposal() function will return the ProposalResponse and a nil error.
For case (2), most error paths return both a newly generated status 500 response and the error.  There is one path though (the one that invokes the simulateProposal) that just returns a status 500 response, but a nil error.
This means that client applications (via SDKs or Gateway) cannot destinguish between chaincode raised errors (case 1) and infrastructure errors (case 2).  This is a problem for the gateway which needs to destinguish these two situations in its retry logic.

This commit fixes this by returning the original error (instead of nil) as well as the generated ProposalResponse for this error path.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>