Create HashedIndex in Pvtdata Store retroactively
This commit add functions in the pvtdata store package that would be used
for creating the HashedIndex in Pvtdata Store retroactively.

In addition, we use this opportunity to upgrade the data format from
V11 to V12, so we can simplify the reading code in the future releases.

Small refactoring is included to simplify the usage of update batch.

Signed-off-by: manish <manish.sethi@gmail.com>