Fix channel config callback in gateway
Currently, the config update callback triggers a discovery call to get the updated channel orderers.
However, this relies on the discovery service having processed the config message before the gateway.  If it hasn’t, the gateway just gets the stale orderer config.
This commit changes the logic so the gateway extracts the orderer endpoints directly from the config bundle.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>