Gateway to use simplified createTx function
The gateway has been reusing the protoutil.CreateTx function to build the transaction envelope from the set of endorsements.  This function changed recently to add extra information into the error message if the proposal responses don’t all match.  This change was identified as causing a problem for the gateway (https://github.com/hyperledger/fabric-gateway/issues/332).
This commit reverts that change, and instead writes the base64 response payloads to the log.  It also refactors the functions slightly to remove duplicate unmarshalling of protos when invoked from the gateway.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>