Update deploy_chaincode.md `package` command docs
Define the default behavior of `peer lifecycle chaincode package`, which does *not* include node_modules in the resulting tar.gz package for chaincode. The existing step of `npm install` and quote "If the command is successful, the JavaScript packages will be installed inside a `node_modules` folder. Now that we that have our dependencies, we can create the chaincode package." would lead a reader to believe that `node_modules` are included when they are not. 

Also added more detail about what chaincode services allow `node_modules` when packaged manually.