Fix gossip unit test flake (#3215)
Increasing the timeout in the gossip service unit tests appears to reduce the occurence of the flake that we frequently observe.

This should be seen as a tactical fix while the underlying cause is investigated.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>
(cherry picked from commit ab3d74183ffcda347f2c09c3dc80f3dd4f946f62)