ledgerutil: Add 'identifytxs' subcommand
This patch adds a new subcommand to ledgerutil that traverses
transactions in a specified ledger and outputs transaction information
(write sets) as a JSON file.
The information is intended to be used to identify a causal transaction
when two ledgers of the same channel are found to be different.

See #2979

Signed-off-by: Taku Shimosawa <taku.shimosawa.uf@hitachi.com>