Update go.mod version to 1.17
The go version directive in go.mod indicates the minimum version of Go that the module works with.
Since Fabric is no longer tested with Go 1.14, it is not known whether Go 1.14 would still
be compatible with current Fabric.
Therefore this commit updates the go directive to 1.17, which is the current version
of Go that Fabric is tested against.

Keeping the go directive up to date with the tested version of go will also
eliminate confusion from users about the recommended version of go.

Note that with the 1.17 version, go.mod has a new format and includes all
indirect depenendencies in a second require stanza.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>