Refactor of ChaincodeEvents service implementation to support resume (#3283)
Enables the client to (optionally) specify an AfterTransactionId property in addition to a start block number when requesting chaincode events, which causes chaincode events up to that transaction ID (inclusive) to be ignored and not returned to the client. This supports resume of chaincode event listening on client reconnect without duplicating or missing any events.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>