Dereference *ed25519.PublicKey and *ed25519.PrivateKey
This is an implementation correction. Since the methods
to store keys deals with interfaces, the ed25519 pointers
were not causing errors on compilation. However, ed25519
key pointers would cause errors in execution time.

Even though these methods to store keys are not used in
production, they might be one day.

Golang deals with ECDSA keys as pointers but not with
ED25519 and I cannot see a reason for that. However,
I believe that dealing with pointers, when it is possible,
is better for memory performance.

Signed-off-by: Johann Westphall <johannwestphall@gmail.com>