Fix ed25519 signature
Previously, the golang ed25519 library was being used to
sign hashes. However, the golang ed25519 library must have
the full message as parameter and it calculates the hash
in the "ed25519.Sign()" function.

To solve that problem, the Sign and Verify methods in
"msp/identities.go" calculate the hash only if the
certificate does not have a ed25519 key.

Some parameters' names were also changed to keep the
coherence with the ed25519 Sign and Verify methods.

Signed-off-by: Johann Westphall <johannwestphall@gmail.com>