Fix fileks and keys for ed25519
Changes regarding dealing with pointers to
ed25519 keys versus the values. This is a
concern because golang generally deals
with ed25519 keys by value.

This causes a difference between the default
way to deal with ecdsa keys (pointers) and
ed25519 ones (values).

As I implemented, the ed25519 keys storing and parsing
follow the general rules:

1) Ed25519 keys are parsed from files as values (ed25519.PrivateKey or
ed25519.PublicKey) but are eventually converted to pointers
to fit the structs ed25519PrivateKey and ed25519PublicKey.

2) To store ed25519 keys in the fileks, they are dealt with as pointers
until the marshalling method, from the x509 golang library. This fits
test cases for storing "nil" keys, that would cause panic if the key
were treated as value.

Fileks test cases for ed25519 were added.

Signed-off-by: Johann Westphall <johannwestphall@gmail.com>