Remove system channel usage from integration tests: discovery
Discovery tests where modified such that they do not use the system channel.

The genesis block is produced either with or without anchor peers, such that the respective tests continue to test the exact functionality.

A the nwo.Network.AnchorsInOrg was modified to allow the creation of a genesis block without anchors and to return all the anchors in an Org, not just one.

A new method nwo.Network.UpdateOrgAnchorPeers was created, to avoid using the deprecated `configtxgen with -outputAnchorPeersUpdate` flag.
The nwo.Network.UpdateOrgAnchorPeers uses an explicit config update instead of `configtxgen`.

A new test case was added to test the explicit re-config of anchor peers using nwo.Network.UpdateOrgAnchorPeers.

AnchorsInOrg
Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: Ic33406572a0cd74789f6e8e61dcaa10f986f7489