Fix approve for my org to not endorse committed transaction
When a chaincode is approved and committed, it means that it has been
accepted and added to the network. However, it is still possible to
approve the exact chaincode definition with the same sequence after it
has been committed. This can lead to confusing behavior because the
transaction is endorsed even though it will never be considered valid.

This commit proposes a fix to prevent the endorsement of transactions
that attempt to approve already committed chaincode definitions with the
same sequence. By failing endorsement of these transactions, we can
provide a better experience for end users. In addition, this will help
to ensure that the chaincode approval process is more transparent and
predictable for all parties involved.

Signed-off-by: Artem Barger <artem@bargr.net>