Gateway registry should provide cluster size (#4000)
When submitting a transaction to BFT orderers, the gateway needs to know the number of consenters in the channel config as well as the array of connected orderers.  The quorum size size is calculated from the former which might be larger than the size of the latter.  These two values should be returned by the same registry lookup to ensure they are not out of sync.

Resolves the PR comment for the previous commit:
https://github.com/hyperledger/fabric/pull/3975#discussion_r1098447996

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>